feat(Collapse): sfc changed to tsx
